upgrade revision number to 8100
This change is to make sure we never try to upgrade config
files written by our version. 8100 is a good round number
greater than the last revision in the original svn.

Also I decided to hardcode it in the .cpp file and not to
pass it from the build system. For git the revision number
make little sense and the only usage of this number is to
upgrade config files. So I tend to think about it as just
an implementation detail of config reading.